+++ Runtime/ObjPool/PoolMember.cs
@@ -11,13 +11,16 @@ namespace OcUtility
         public UnityEvent OnWakeUp;
         public UnityEvent OnSleep;
         public OcPool Pool { get; set; }
-        public void WakeUp()
+        
+        /// <summary> 오브젝트를 활성화하고 OnWakeUp 콜백을 실행함. OnEnable에서 호출하지 말 것. </summary>
+        public virtual void WakeUp()
         {
             gameObject.SetActive(true);
             OnWakeUp?.Invoke();
         }
 
-        public void Sleep()
+        /// <summary> OnSleep 콜백을 실행하고 오브젝트를 비활성화한 후, 원래의 풀로 되돌아감. OnDisable에서 호출하지 말 것. </summary>
+        public virtual void Sleep()
         {
             OnSleep?.Invoke();
             gameObject.SetActive(false);